fix: implement timer preemption and fix SENDNB/RS notification storm
**Timer preemption:** `proc_no_time` sets `RTS_PREEMPTED` for
kernel-scheduled processes. The syscall handler re-enqueues at tail,
giving
other processes at the same priority a chance to run. Without this,
a runnable process at the head of its queue starved all others.

**Server replies use SENDNB (non-blocking):** PM, VM, and RS all used
blocking `SEND` (syscall 46) to reply. If the destination wasn't
receiving,
the sender blocked with `RTS_SENDING` forever — locking the IPC system.
Changed all three to `SENDNB` (syscall 51). Also added the
`SYS_IPC_SENDNB`
kernel handler at syscall 51.

**RS→PM notification storm:** RS was the source of ~36K
notifications/sec
to PM. RS's blocking `SEND` would wedge, then RS's main loop behavior
generated continuous `mini_notify(RS, PM)` calls. Fixed by RS using
`SENDNB`.

**send_sig notification guard:** `send_sig` always called
`mini_notify(SYSTEM, target)` even when the target already had
`RTS_SIGNALED`. Added C's `!RTS_ISSET(rp, RTS_SIGNALED)` guard.
Without this, every timer tick flooded PM with notifications.

**SYS_ENDKSIG after SYS_GETKSIG:** PM called `get_work()` (kernel call
7)
to collect exited children but never called `end_work()` (kernel call
8).
Without it, `SIG_PENDING` accumulated in the target's rts_flags.

**p_pending type fix:** `p_pending` was `u32` (4 bytes) but C MINIX's
`sigset_t` is 16 bytes (`__uint32_t __bits[4]`). `do_getksig_handler`
copied 16 bytes via `copy_from_slice` — the length mismatch panicked,
and the panic handler (`hlt_loop`) silently halted.

**IPC status layout fix:** `ipc_status_add_call`,
`ipc_status_add_flags`,
`ipc_status_clear` now match C layout: call type in low 6 bits of
`p_misc_flags`, flags shifted to bits 16+.

**syscall register clobbers:** All `syscallN` inline asm blocks declare
`lateout` clobbers for rdi, rsi, rdx, r8, r9, r10, r12-r15. Missing
clobbers let LLVM hold stale values across the syscall instruction.

**Boot process priorities:** VM and RS get priority 7 / 200ms quantum;
PM, VFS, MFS, INIT get priority 0 / 50ms quantum — matching C MINIX.
Child processes inherit parent priority instead of hardcoded 5.

refactor: remove kernel serial traces and fix signal loop / shell errors

Remove verbose serial tracing from kernel IPC, scheduling, and syscall
paths that was used for early debugging. The traces were polluting the
serial output and are no longer needed.

- kernel-boot: remove syscall handler and message delivery traces
- kernel/ipc: remove mini_send, do_sync_ipc, and delivermsg traces
- kernel/sched: remove enqueue trace
- kernel/syscall: remove sys_ipc_send_handler trace
- kernel/system: remove do_getksig_handler trace

Fixes found during cleanup:
- pm: limit signal loop to 50 iterations to prevent hangs when signals
  regenerate faster than they can be drained
- pm: remove hex endpoint debug writes
- userland/sh: save original command string at offset 200 in cmd_path
  to survive waitpid overwriting bytes 0-63, improving "not found"
  error messages with proper quoting
- minix-rt: add #[inline(never)] to waitpid

+++ crates/kernel/src/proc.rs
@@ -294,7 +294,8 @@ pub struct Proc {
     pub p_sendto_e: i32,
 
     /// Bit map for pending kernel signals.
-    pub p_pending: u32,
+    /// Matches C MINIX `sigset_t` (16 bytes on NetBSD/Minix).
+    pub p_pending: u128,
 
     /// Process name (null-terminated).
     pub p_name: [u8; PROC_NAME_LEN],
@@ -362,7 +363,7 @@ impl Default for Proc {
             p_q_link: core::ptr::null_mut(),
             p_getfrom_e: 0,
             p_sendto_e: 0,
-            p_pending: 0,
+            p_pending: 0u128,
             p_name: [0u8; PROC_NAME_LEN],
             p_endpoint: 0,
             p_sendmsg: [0u8; MESSAGE_SIZE],
